refactor UserControllerTest and MemberControllerTest
