+++ src/test/java/com/alkemy/ong/web/controllers/UserControllerTest.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong;
+package com.alkemy.ong.web.controllers;
 
 import com.alkemy.ong.data.entities.RoleEntity;
 import com.alkemy.ong.data.entities.UserEntity;
